+++ changelog.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## v0.4.0 (2017-01-04)
+
+- Merged `ServiceLocator` and `ServiceRegister` into a new interface `ServiceRegistry`
+- Added trace options for camel routes in channels
+- Telegram content formatter
+- Link expression changed to `title|url` instead of `title <url>`
+
 ## v0.3.0 (2016-12-16)
 
 - Subscription commands simplified (`subscribe`, `unsubscribe` and `subscriptions`)